# Implements a performance tracker