--- a/src/AddressExtractorMonitor.cs
+++ b/src/AddressExtractorMonitor.cs
@@ -1,9 +1,12 @@
 using System.Collections.Concurrent;
 using System.Diagnostics;
+using MyAddressExtractor.Objects.Performance;
 
 namespace MyAddressExtractor {
     public class AddressExtractorMonitor : IAsyncDisposable {
         private readonly AddressExtractor Extractor = new();
+        private readonly IPerformanceStack Stack;
+
         protected readonly IDictionary<string, Count> Files = new ConcurrentDictionary<string, Count>();
         protected readonly ISet<string> Addresses = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
         protected int Lines { get; private set; }
@@ -11,9 +14,10 @@ namespace MyAddressExtractor {
         protected readonly Stopwatch Stopwatch = Stopwatch.StartNew();
         private readonly Timer Timer;
 
-        public AddressExtractorMonitor(): this(TimeSpan.FromMinutes(1)) {}
-        public AddressExtractorMonitor(TimeSpan iterate)
+        public AddressExtractorMonitor(IPerformanceStack stack): this(stack, TimeSpan.FromMinutes(1)) {}
+        public AddressExtractorMonitor(IPerformanceStack stack, TimeSpan iterate)
         {
+            this.Stack = stack;
             this.Timer = new Timer(_ => this.Log(), null, iterate, iterate);
         }
 
@@ -21,19 +25,22 @@ namespace MyAddressExtractor {
         {
             foreach (var inputFilePath in files)
             {
-                var count = new Count();
-                var addresses = 0;
-
-                this.Files.Add(inputFilePath, count);
-
-                var parser = FileExtensionParsing.GetFromPath(inputFilePath);
-                await using (var reader = parser.GetReader(inputFilePath))
+                using (var stack = this.Stack.CreateStack("Read file"))
                 {
-                    await foreach(var email in this.Extractor.ExtractAddressesAsync(reader, cancellation))
+                    var count = new Count();
+                    var addresses = 0;
+
+                    this.Files.Add(inputFilePath, count);
+
+                    var parser = FileExtensionParsing.GetFromPath(inputFilePath);
+                    await using (var reader = parser.GetReader(inputFilePath))
                     {
-                        if (this.Addresses.Add(email))
-                            count.Value = addresses++;
-                        this.Lines++;
+                        await foreach(var email in this.Extractor.ExtractAddressesAsync(stack, reader, cancellation))
+                        {
+                            if (this.Addresses.Add(email))
+                                count.Value = addresses++;
+                            this.Lines++;
+                        }
                     }
                 }
             }
@@ -46,6 +53,8 @@ namespace MyAddressExtractor {
             long rate = (long)(this.Lines / (this.Stopwatch.ElapsedMilliseconds / 1000.0));
             Console.WriteLine($"Read lines total: {this.Lines:n0}");
             Console.WriteLine($"Read lines rate: {rate:n0}/s\n");
+
+            this.Stack.Log();
         }
 
         internal async ValueTask SaveAsync(CommandLineProcessor cli, CancellationToken cancellation = default)