--- a/src/Extensions.cs
+++ b/src/Extensions.cs
@@ -1,3 +1,5 @@
+using System.Diagnostics;
+
 namespace MyAddressExtractor {
     public static class Extensions {
         public static void AddAll<TKey, TVal>(this IDictionary<TKey, TVal> dictionary, IEnumerable<TKey> keys, TVal value)
@@ -6,6 +8,13 @@ namespace MyAddressExtractor {
             {
                 dictionary[key] = value;
             }
+        }
+
+        public static TimeSpan GetAndReset(this Stopwatch stopwatch)
+        {
+            var millis = stopwatch.Elapsed;
+            stopwatch.Restart();
+            return millis;
         }
     }
 }