--- a/src/Objects/Performance/IPerformanceStack.cs
+++ b/src/Objects/Performance/IPerformanceStack.cs
@@ -5,8 +5,14 @@ namespace MyAddressExtractor.Objects.Performance {
     public interface IPerformanceStack : IDisposable {
         static readonly IPerformanceStack DEFAULT = new DefaultPerformanceStack();
 
+        /// <summary>
+        /// Create a new nested stack
+        /// </summary>
         IPerformanceStack CreateStack(string name);
 
+        /// <summary>
+        /// Add a new measured step in the current stack
+        /// </summary>
         void Step(string name);
 
         void Log();