@GStefanowich 2023-04-29
@troyhunt Merge conflict resolved if you want to merge

You can do some indepth testing on what parts of the code are slowest since you can throw some data leaks at it